was starting to see floating point problems
the engine was initially designed to be able to show where satellites are in
real time (as in "now"). however, with time acceleration, and no particular
attention to accuracy, that stuff is rather useless.

however, the large time offset from "epoch" was starting to make the orbit
motion jerky due to floating point accuracy.